requirements.txt
pyquery
requests
sqlalchemy
ijson
werkzeug==2.3.0
flask
flask-login
flask-sqlalchemy
flask-migrate
flask_socketio
flask-wtf
mysql-connector-python
email_validator
pyquery
pyopenssl
requests
flask_assets
cssmin